pyproject.toml
[tool.poetry]
name = "nonebot_plugin_withdraw"
version = "0.4.0"
description = "A simple withdraw plugin for Nonebot2"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-withdraw"
repository = "https://github.com/noneplugin/nonebot-plugin-withdraw"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.0"
nonebot-plugin-alconna = ">=0.51.4"
nonebot-plugin-uninfo = ">=0.4.0,<1.0.0"

[tool.poetry.group.adapters]
optional = true

[tool.poetry.group.adapters.dependencies]
nonebot-adapter-onebot = "^2.4.4"
nonebot-adapter-kaiheila = "^0.3.4"
nonebot-adapter-telegram = "^0.1.0b17"
nonebot-adapter-feishu = "^2.6.2"
nonebot-adapter-red = "^0.9.0"
nonebot-adapter-discord = "^0.1.8"
nonebot-adapter-satori = "^0.12.3"
nonebot-adapter-qq = "^1.5.0"
nonebot-adapter-dodo = "^0.2.1"

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
